Speed up heartbeat sending by 2s
Substract two seconds, because some devices (like Elli Connect/Pro) with OPEV/OSCEV interpret the heartbeat timeout (<= 4s) as the time within which a heartbeat should be received and otherwise will go into fallback mode.
But other EVSE devices and in LPC (<= 60s), the heartbeat should be considered missing, if it is not received within twice the heartbeat timeout timeframe.