Add support for handling duplicate read messages
If different use case implementations trigger the same read messages, currently they are all sent through and result in multiple responses.

This change will cache up to the last 20 read messages for each feature and not resend duplicates to the same remote feature. Instead it will return the message counter of the previous read request, so all requestors will get the same responses.