Add support for writing non partial data
When the remote service does not provide support for partial writes, it is required that all data is always written. Which means that the changed data has to be merged with the currently known remote data, or that the full data is provided in the write in the first place.

This commit adds support for a merge, if merge support is implemented. Otherwise it will return an error, and the data will be sent as is.