Fix: Correct Data Types for Setpoint Description, Selector, and HVAC
This commit addresses the data type issues in the `Setpoint` and `HVAC` model messages, specifically
for `SetpointDescriptionDataType`, `SetpointDescriptionListDataSelectorsType`, and
`HvacSystemFunctionSetpointRelationDataType`. It also corrects invalid types in `function_data_factory.go`
for `FunctionTypeHvacOperationModeDescriptionListData` and `FunctionTypeHvacSystemFunctionDescriptionListData`.

Changes include:
1. Updating `SetpointId` in `hvacSystemFunctionSetpointRelationListData` to be a list, as specified in the EEBus
   SPINE Technical Specification Resource Specification.
2. Correcting data types for fields in `SetpointDescriptionDataType` and `SetpointDescriptionListDataSelectorsType`
   also according to the Resource Specification.

Before this commit, sending `FunctionTypeHvacOperationModeDescriptionListData` and
`FunctionTypeHvacSystemFunctionOperationModeRelationListData` failed due to invalid
data types provided to `createFunctionData` in the factory.

Additionally, the `SetpointId` field needed to be a list, not a single value.  According to the specification, an
operation mode can have multiple setpoints (up to four), such as for the "auto" operation mode.

Sending `FunctionTypeSetpointDescriptionListData` also failed due to incorrect field data types.

With these fixes, requests for `FunctionTypeSetpointDescriptionListData`, `FunctionTypeHvacOperationModeDescriptionListData`,
and `FunctionTypeHvacSystemFunctionOperationModeRelationListData` now work correctly.