Update handling of detaileddiscovery updates
Previously if multiple entities where added in an update, addition was happening also multiple times for every entity. Also a removed entity was “added” as well. The code checked if the entity already existed and exited, so there was no issue with this. But it was unnecessary to do so.

This change adds the option to only add a specific entity of a dataset to improve this.