Improve device change handling
Instead of using the `message.FeatureRemote` in the DeviceChange message, change the logic in device_local.

When a remote device is adding, we need to send a subscription request to the NodeManagement feature. As the remote device, the entity address, feature type and role are all known, fetch the feature and its address.