Improved Unmarshalling
SHIPs previous simple removing of JSON list elements caused empty arrays to be converted to {} being an empty object. This caused unmarshalling issues.

This change now fixes this by using typed unmarshalling where we evaluate every instance of {} and check if it is an array type or object and then convert it properly.